// sdks/remote_plugin/nodejs/src/proto/plugin.ts
export enum DropStrategy {
  DROP_STRATEGY_UNSPECIFIED = 0,
  DROP_STRATEGY_BLOCK_IF_FULL = 1,
  DROP_STRATEGY_DROP_NEWEST = 2,
  DROP_STRATEGY_DROP_OLDEST = 3,
  UNRECOGNIZED = -1,
}
// ...
export enum EventType {
  EVENT_TYPE_UNSPECIFIED = 0,
  EVENT_TYPE_SIGNAL = 1,
  EVENT_TYPE_PAYLOAD = 2,
  EVENT_TYPE_AUDIO = 3,
  EVENT_TYPE_VIDEO = 4,
  UNRECOGNIZED = -1,
}
// ...
export enum ValueType {
  VALUE_TYPE_UNSPECIFIED = 0,
  VALUE_TYPE_STRING = 1,
  VALUE_TYPE_NUMBER = 2,
  VALUE_TYPE_INTEGER = 3,
  VALUE_TYPE_BOOLEAN = 4,
  VALUE_TYPE_OBJECT = 5,
  VALUE_TYPE_ARRAY = 6,
  UNRECOGNIZED = -1,
}
// ...
export interface PluginMetadata {
  name: string;
  description: string;
  schema: string;
  inputs: Property[];
  outputs: Property[];
  ports: PortMetadata[];
  multiTrack: MultiTrackConfig | undefined;
}

export interface MultiTrackConfig {
  enabled: boolean;
  payload: TrackConfig | undefined;
}

export interface TrackConfig {
  enabled: boolean;
  bufferSize: number;
  dropStrategy: DropStrategy;
  interruptSignals: string[];
}

export interface Property {
  type: EventType;
  name: string;
  fields: Field[];
}

export interface Field {
  key: string;
  type: ValueType;
  required: boolean;
}

export interface PortMetadata {
  type: EventType;
  port: number;
  name: string;
  description: string;
}
// ...
export const MultiTrackConfig: MessageFns<MultiTrackConfig> = {
  encode(message: MultiTrackConfig, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.enabled !== false) {
      writer.uint32(8).bool(message.enabled);
    }
    if (message.payload !== undefined) {
      TrackConfig.encode(message.payload, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MultiTrackConfig {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMultiTrackConfig();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.enabled = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.payload = TrackConfig.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MultiTrackConfig {
    return {
      enabled: isSet(object.enabled) ? globalThis.Boolean(object.enabled) : false,
      payload: isSet(object.payload) ? TrackConfig.fromJSON(object.payload) : undefined,
    };
  },

  toJSON(message: MultiTrackConfig): unknown {
    const obj: any = {};
    if (message.enabled !== false) {
      obj.enabled = message.enabled;
    }
    if (message.payload !== undefined) {
      obj.payload = TrackConfig.toJSON(message.payload);
    }
    return obj;
  },

  create(base?: DeepPartial<MultiTrackConfig>): MultiTrackConfig {
    return MultiTrackConfig.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<MultiTrackConfig>): MultiTrackConfig {
    const message = createBaseMultiTrackConfig();
    message.enabled = object.enabled ?? false;
    message.payload = (object.payload !== undefined && object.payload !== null)
      ? TrackConfig.fromPartial(object.payload)
      : undefined;
    return message;
  },
};
// ...
export const Property: MessageFns<Property> = {
  encode(message: Property, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    for (const v of message.fields) {
      Field.encode(v!, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Property {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseProperty();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.fields.push(Field.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Property {
    return {
      type: isSet(object.type) ? eventTypeFromJSON(object.type) : 0,
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      fields: globalThis.Array.isArray(object?.fields) ? object.fields.map((e: any) => Field.fromJSON(e)) : [],
    };
  },

  toJSON(message: Property): unknown {
    const obj: any = {};
    if (message.type !== 0) {
      obj.type = eventTypeToJSON(message.type);
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.fields?.length) {
      obj.fields = message.fields.map((e) => Field.toJSON(e));
    }
    return obj;
  },

  create(base?: DeepPartial<Property>): Property {
    return Property.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<Property>): Property {
    const message = createBaseProperty();
    message.type = object.type ?? 0;
    message.name = object.name ?? "";
    message.fields = object.fields?.map((e) => Field.fromPartial(e)) || [];
    return message;
  },
};

function createBaseField(): Field {
  return { key: "", type: 0, required: false };
}

export const Field: MessageFns<Field> = {
  encode(message: Field, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.key !== "") {
      writer.uint32(10).string(message.key);
    }
    if (message.type !== 0) {
      writer.uint32(16).int32(message.type);
    }
    if (message.required !== false) {
      writer.uint32(24).bool(message.required);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Field {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseField();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.key = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.required = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Field {
    return {
      key: isSet(object.key) ? globalThis.String(object.key) : "",
      type: isSet(object.type) ? valueTypeFromJSON(object.type) : 0,
      required: isSet(object.required) ? globalThis.Boolean(object.required) : false,
    };
  },

  toJSON(message: Field): unknown {
    const obj: any = {};
    if (message.key !== "") {
      obj.key = message.key;
    }
    if (message.type !== 0) {
      obj.type = valueTypeToJSON(message.type);
    }
    if (message.required !== false) {
      obj.required = message.required;
    }
    return obj;
  },

  create(base?: DeepPartial<Field>): Field {
    return Field.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<Field>): Field {
    const message = createBaseField();
    message.key = object.key ?? "";
    message.type = object.type ?? 0;
    message.required = object.required ?? false;
    return message;
  },
};

function createBasePortMetadata(): PortMetadata {
  return { type: 0, port: 0, name: "", description: "" };
}

export const PortMetadata: MessageFns<PortMetadata> = {
  encode(message: PortMetadata, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    if (message.port !== 0) {
      writer.uint32(16).int32(message.port);
    }
    if (message.name !== "") {
      writer.uint32(26).string(message.name);
    }
    if (message.description !== "") {
      writer.uint32(34).string(message.description);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PortMetadata {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePortMetadata();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.port = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.description = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PortMetadata {
    return {
      type: isSet(object.type) ? eventTypeFromJSON(object.type) : 0,
      port: isSet(object.port) ? globalThis.Number(object.port) : 0,
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      description: isSet(object.description) ? globalThis.String(object.description) : "",
    };
  },

  toJSON(message: PortMetadata): unknown {
    const obj: any = {};
    if (message.type !== 0) {
      obj.type = eventTypeToJSON(message.type);
    }
    if (message.port !== 0) {
      obj.port = Math.round(message.port);
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.description !== "") {
      obj.description = message.description;
    }
    return obj;
  },

  create(base?: DeepPartial<PortMetadata>): PortMetadata {
    return PortMetadata.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<PortMetadata>): PortMetadata {
    const message = createBasePortMetadata();
    message.type = object.type ?? 0;
    message.port = object.port ?? 0;
    message.name = object.name ?? "";
    message.description = object.description ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}